Move files to the right packages
